Move restTemplateFactory to service package.
