+++ backend/src/main/java/ch/zuehlke/fullstack/hackathon/service/InsightClient.java
@@ -1,10 +1,10 @@
 package ch.zuehlke.fullstack.hackathon.service;
 
-import ch.zuehlke.fullstack.hackathon.RestTemplateFactory;
 import ch.zuehlke.fullstack.hackathon.model.UserInfo;
 import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.util.UriComponentsBuilder;
